requirements.txt
streamlit
python-dotenv
google-genai
pymongo
requests
datetime